corregidos modelos (se resetearon las tablas) / corregido manejo de stock / corregido manejo de excepciones (se incluye manejo global para evitar los try-catch en los controladores)

+++ src/main/java/ar/com/tresimplesrazones/tsr/model/Compra.java
@@ -16,7 +16,7 @@ public class Compra {
     @Column(name = "cantidad_comprada")
     private int cantidadComprada;
     @Column(name = "precio_de_compra")
-    private int precioDeCompra;
+    private Long precioDeCompra;
     @Column(name = "fecha_de_compra", columnDefinition = "DATE")
     private LocalDate fechaDeCompra;
